Improved Bower support
1) Bower actions are not blocking the UI anymore
2) Fix an issue with ranges
3) Fix an issue while reading bower.json files

+++ bundles/org.eclipse.wst.jsdt.bower.ide.ui/src/main/java/org/eclipse/wst/jsdt/bower/ide/ui/internal/handlers/BowerClearHandler.java
@@ -13,6 +13,11 @@ package org.eclipse.wst.jsdt.bower.ide.ui.internal.handlers;
 import org.eclipse.core.resources.IFolder;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.IStatus;
+import org.eclipse.core.runtime.OperationCanceledException;
+import org.eclipse.core.runtime.Status;
+import org.eclipse.wst.jsdt.bower.ide.ui.internal.utils.I18n;
+import org.eclipse.wst.jsdt.bower.ide.ui.internal.utils.I18nKeys;
 import org.eclipse.wst.jsdt.bower.ide.ui.internal.utils.IBowerIdeUiConstants;
 import org.eclipse.wst.jsdt.nodejs.core.api.utils.ILogger;
 
@@ -29,16 +34,35 @@ public class BowerClearHandler extends AbstractBowerHandler {
 	 * @see org.eclipse.wst.jsdt.bower.ide.ui.internal.handlers.AbstractBowerHandler#doExecute(org.eclipse.core.runtime.IProgressMonitor)
 	 */
 	@Override
-	protected void doExecute(IProgressMonitor monitor) {
+	protected IStatus doExecute(IProgressMonitor monitor) {
+		IStatus status = Status.OK_STATUS;
 		if (this.getBowerJson().isPresent()) {
 			IFolder outputDirectory = this.getOutputDirectory();
 			if (outputDirectory != null && outputDirectory.exists()) {
-				try {
-					outputDirectory.delete(true, monitor);
-				} catch (CoreException e) {
-					logger.log(IBowerIdeUiConstants.BOWER_IDE_UI_BUNDLE_ID, ILogger.ERROR, e);
+				if (!monitor.isCanceled()) {
+					try {
+						outputDirectory.delete(true, monitor);
+					} catch (OperationCanceledException e) {
+						// do not log
+						status = Status.CANCEL_STATUS;
+					} catch (CoreException e) {
+						logger.log(IBowerIdeUiConstants.BOWER_IDE_UI_BUNDLE_ID, ILogger.ERROR, e);
+					}
+				} else {
+					status = Status.CANCEL_STATUS;
 				}
 			}
 		}
+		return status;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 *
+	 * @see org.eclipse.wst.jsdt.bower.ide.ui.internal.handlers.AbstractBowerHandler#getJobName()
+	 */
+	@Override
+	protected String getJobName() {
+		return I18n.getString(I18nKeys.BOWER_CLEAR_JOB_NAME);
 	}
 }
